requirements.txt
playwright>=1.40.0
playwright-stealth>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
apscheduler>=3.10.0
click>=8.1.0
plyer>=2.1.0
rich>=13.0.0
